debug

+++ test/lumen_spec_sens_test/load_datasets.jl
@@ -5,7 +5,7 @@ using DataFrames, CSV, Serialization
 
 # datatreatments parameters
 rng = 42
-float_type = Float32
+float_type = Float64
 
 # solexplorer parameters
 models = Dict(
